Messages creation passes tests

+++ src/main/java/com/example/repository/AccountRepository.java
@@ -6,6 +6,6 @@ import com.example.entity.Account;
 import java.util.Optional;
 
 
-public interface AccountRepository extends JpaRepository<Account, Long>{
+public interface AccountRepository extends JpaRepository<Account, Integer>{
     Optional<Account> findByUsername(String username);
 }
